Implement new abstract lephare functions (#27)
* Started work on updating rail to reflect abstracted classes

We would liek the option to run autoadapt either on the inform training set or on each chunk

* Updated params to take lephare config object

* Implemented abstract prepare and process methods

Still to implement autoadapt in inform stage. We need to convert the config key map to a dictionary for it to be pickled. It can easily be converted back
by looping over the keys and replacing the values with lephare.keyword objects

* Implementing offsets in inform stage

* Undo notebook changes

* Add run naming functionality

This depends on the changes to DataManager in lephare being merged and published. There may be situations where the user wants to default to imed runs but currently they have to do this via the name of the inform stage

---------

Co-authored-by: Raphael Shirley <rshirley@mpe.mpg.de>