Fixed model error and changed code to set run without lin
We should perhaps set the link to the run directory in a more robust way. currently I get it from the default work directory and go up a level to runs.
We also might wnat to make it overwritable in the estimator so that the folder can be moved between systems